# optimizer: inline union-split, abstract callsite