@aviatesk 2021-11-30
What do you mean by unlinking threshold?