@aviatesk 2021-11-30
Okay, and yeah, probably. 
Or we may want to enable this inlining analysis only when some flag is enabled, since it seems like it leads to bloated code and hence worse performance if applied globally.
Something like:
```julia
@inline :aggressive isType(t::DataType) = t.name === Type.body.name
@inline :aggressive isType(@nospecialize t) = false

# or

let x::Any
    @inline :aggressive isType(x)
end
```