@aviatesk 2022-03-03
@nanosoldier `runbenchmarks(!"scalar", vs=":master")`