@nanosoldier 2022-03-03
[Your benchmark job](https://github.com/JuliaLang/julia/pull/42809#issuecomment-1058238690) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/e849d6b_vs_82dc130/report.md).