@aviatesk 2022-03-07
@nanosoldier `runbenchmarks(!"scalar", vs=":master")`