@nanosoldier 2022-03-07
[Your benchmark job](https://github.com/JuliaLang/julia/pull/42809#issuecomment-1060244483) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/08573a3_vs_811e534/report.md).